Feat: add prop destructuring to post / post card
