+++ src/client/components/Post/types.ts
@@ -1,0 +1,17 @@
+interface PostComment {
+  body: string;
+  author: string;
+}
+
+export interface PostProps {
+  content: string;
+  image: string;
+  title: string;
+  comments?: PostComment[];
+  comment: {
+    onSubmit: Function;
+    onChange: Function;
+    input: string;
+    onKeyUp: Function;
+  }
+}
