+++ src/client/components/PostCard/types.ts
@@ -1,0 +1,6 @@
+export interface PostCardProps {
+  content: string;
+  id: string;
+  image: string;
+  title: string;
+}
